Improve logic for adding new resources.
- use helpers instead of relying on blank example in vfs
- don't generate invalid/duplicate names
